Fix File IO

+++ lib/bloc/data/data_bloc.dart
@@ -325,30 +325,15 @@ class DataBloc extends Bloc<DataEvent, DataState> {
 // ** UserLoadFile Event **
 // ************************
   Stream<DataState> _mapUserLoadFileState(UserLoadFile event) async* {
-    // Check if Path exists
-    bool exists = await Directory(event.file.metadata.path).exists();
+    // Get Bytes
+    Uint8List bytes = await event.file.raw.readAsBytes();
 
-    // Return Success
-    if (exists) {
-      // Get Bytes
-      Uint8List bytes = await event.file.raw.readAsBytes();
-
+    // Check Bytes
+    if (bytes != null) {
       // Change State
       yield UserViewingFileSuccess(bytes, event.file.metadata);
-    }
-    // Return Failure
-    else {
-      // Load Provider
-      MetadataProvider metadataProvider = new MetadataProvider();
-      await metadataProvider.open();
-
-      // Remove from DB
-      await metadataProvider.delete(event.file.metadata.id);
-
-      // Reload Files
-      add(UserGetAllFiles());
-
-      // Change State
+    }else{
+      // Send Failure
       yield UserViewingFileFailure();
     }
   }
@@ -365,9 +350,15 @@ class DataBloc extends Bloc<DataEvent, DataState> {
     int count = await metadataProvider.delete(event.meta.id);
     log.i(count.toString() + "Files Removed from database");
 
-    // Remove from LocalData
-    var ref = File(event.meta.path);
-    await ref.delete();
+    // Check if Path exists
+    bool exists = await Directory(event.meta.path).exists();
+
+    // File Exists
+    if (exists) {
+      // Remove from LocalData
+      var ref = File(event.meta.path);
+      await ref.delete();
+    }
 
     // Reload Files
     add(UserGetAllFiles());
